refactor(KERN): remove unused ArraySummary components
