+++ app/components/content/arraySummary/KernArraySummaryDataItem.tsx
