+++ app/components/content/arraySummary/__test__/ArraySummaryDataItems.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
+import KernArraySummaryDataItems from "~/components/content/arraySummary/KernArraySummaryDataItem";
 import type { ArrayConfigClient } from "~/services/array";
-import ArraySummaryDataItems from "../ArraySummaryDataItems";
 
 const mockArrayConfiguration: ArrayConfigClient = {
   event: "add-unterhaltszahlungen",
@@ -15,9 +15,12 @@ const mockDataItem = {
   familyRelationship: "mother",
 };
 
-vi.mock("~/components/content/arraySummary/ArraySummaryItemButton", () => ({
-  default: () => <div>Mock ArraySummaryItemButton</div>,
-}));
+vi.mock(
+  "~/components/content/arraySummary/KernArraySummaryItemActions",
+  () => ({
+    default: () => <div>Mock ArraySummaryItemActions</div>,
+  }),
+);
 
 describe("ArraySummaryDataItems", () => {
   afterEach(() => {
@@ -32,7 +35,7 @@ describe("ArraySummaryDataItems", () => {
       surname: "Nachname",
     };
     const { getByText } = render(
-      <ArraySummaryDataItems
+      <KernArraySummaryDataItems
         configuration={mockArrayConfiguration}
         items={mockDataItem}
         itemIndex={0}
@@ -41,9 +44,7 @@ describe("ArraySummaryDataItems", () => {
       />,
     );
 
-    Object.entries(itemLabels).forEach(([key, label]) => {
-      const expectedTagName = key.includes(".") ? "DIV" : "P";
-      expect(getByText(label).tagName).toBe(expectedTagName);
+    Object.values(itemLabels).forEach((label) => {
       expect(getByText(label)).toBeInTheDocument();
     });
   });
@@ -55,7 +56,7 @@ describe("ArraySummaryDataItems", () => {
     };
 
     const { container } = render(
-      <ArraySummaryDataItems
+      <KernArraySummaryDataItems
         configuration={mockArrayConfiguratinWithHiddenFields}
         items={mockDataItem}
         itemIndex={0}
@@ -69,7 +70,7 @@ describe("ArraySummaryDataItems", () => {
 
   it("should render ArraySummaryDataItems with subtitle", () => {
     const { getByText } = render(
-      <ArraySummaryDataItems
+      <KernArraySummaryDataItems
         configuration={mockArrayConfiguration}
         items={mockDataItem}
         itemIndex={0}
@@ -86,7 +87,7 @@ describe("ArraySummaryDataItems", () => {
     const { queryByText } = render(
       <>
         {[0, 1].map((itemIndex) => (
-          <ArraySummaryDataItems
+          <KernArraySummaryDataItems
             key={itemIndex}
             configuration={mockArrayConfiguration}
             items={mockDataItem}
@@ -107,7 +108,7 @@ describe("ArraySummaryDataItems", () => {
     const { queryByText } = render(
       <>
         {[0, 1].map((itemIndex) => (
-          <ArraySummaryDataItems
+          <KernArraySummaryDataItems
             key={itemIndex}
             configuration={{
               ...mockArrayConfiguration,
